# docs/openapi.yaml
basePath: /api
definitions:
...
  handlers.CreateTeamRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  models.Challenge:
    properties:
      category:
        type: string
      created_at:
        type: string
      decay_threshold:
        type: integer
      description:
        type: string
      difficulty:
        type: string
      dynamic_scoring:
        type: boolean
      file_url:
        description: 'Deprecated: use ChallengeFiles'
        type: string
      id:
        type: string
      initial_points:
        type: integer
      minimum_points:
        type: integer
      name:
        type: string
      sql_dataset_url:
        type: string
      sql_enabled:
        type: boolean
      sql_schema_hint:
        type: string
      tags:
        type: string
      updated_at:
        type: string
      visible:
        type: boolean
    type: object
  models.Competition:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_at:
        type: string
      id:
        type: integer
      name:
        type: string
      registration_end:
        type: string
      registration_start:
        type: string
      rules_html:
        type: string
      scoreboard_blackout:
        type: boolean
      scoreboard_frozen:
        type: boolean
      start_at:
        type: string
      status:
        description: draft|registration|running|ended
        type: string
      updated_at:
        type: string
    type: object
  models.CompetitionScoreboardEntry:
    properties:
      last_solve:
        type: string
      rank:
        type: integer
      score:
        type: integer
      solve_count:
        type: integer
      team_id:
        type: string
      team_name:
        type: string
    type: object
...
  models.Team:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      invite_id:
        type: string
      invite_permission:
        type: string
      name:
        type: string
      owner_id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8090
info:
  contact: {}
  description: Self-hosted CTF platform API. Most write endpoints require authentication
    via JWT cookie (auth_token).
  title: hCTF API
  version: 1.0.0
paths:
...
  /api/admin/competitions:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Competition'
      security:
      - CookieAuth: []
      summary: Create a competition (admin only)
      tags:
      - Competitions
  /api/admin/competitions/{id}:
    delete:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - CookieAuth: []
      summary: Delete a competition (admin only)
      tags:
      - Competitions
    put:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - CookieAuth: []
      summary: Update a competition (admin only)
      tags:
      - Competitions
  /api/admin/competitions/{id}/blackout:
    post:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - CookieAuth: []
      summary: Toggle competition scoreboard blackout (admin only)
      tags:
      - Competitions
  /api/admin/competitions/{id}/challenges:
    post:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - CookieAuth: []
      summary: Add challenge to competition (admin only)
      tags:
      - Competitions
  /api/admin/competitions/{id}/challenges/{cid}:
    delete:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      - description: Challenge ID
        in: path
        name: cid
        required: true
        type: string
      responses: {}
      security:
      - CookieAuth: []
      summary: Remove challenge from competition (admin only)
      tags:
      - Competitions
  /api/admin/competitions/{id}/force-end:
    post:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - CookieAuth: []
      summary: Force-end a competition (admin only)
      tags:
      - Competitions
  /api/admin/competitions/{id}/force-start:
    post:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - CookieAuth: []
      summary: Force-start a competition (admin only)
      tags:
      - Competitions
  /api/admin/competitions/{id}/freeze:
    post:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - CookieAuth: []
      summary: Toggle competition scoreboard freeze (admin only)
      tags:
      - Competitions
  /api/admin/competitions/{id}/teams:
    get:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - CookieAuth: []
      summary: List teams registered for competition
      tags:
      - Competitions
...
  /api/competitions/{id}:
    get:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Competition'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get competition by ID
      tags:
      - Competitions
  /api/competitions/{id}/register:
    post:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Register current user's team for a competition
      tags:
      - Competitions
  /api/competitions/{id}/scoreboard:
    get:
      parameters:
      - description: Competition ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CompetitionScoreboardEntry'
            type: array
      summary: Get competition scoreboard
      tags:
      - Competitions
...
  /challenges:
    get:
      description: Returns all visible challenges. Admins also see hidden challenges.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Challenge'
            type: array
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: List all challenges
      tags:
      - Challenges
...
  /teams:
    get:
      description: Invite codes are only visible to team members.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: List all teams
      tags:
      - Teams
    post:
      consumes:
      - application/json
      parameters:
      - description: Team data
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Team'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create a new team
      tags:
      - Teams
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"